requirements.txt
pytorch
torchvision
torchviz
matplotlib
libpng
jpeg